Ensure route-specific data is cleared if the route changes
It actually clears the route data after it has been used once.
This has the desirable side-effect of not accumulating messages
and errors.

Usually, the data will be saved if the client side validation
fails, and the user is redirected to the route containing the
form they already filled out, with the fields populated with
the saved values. So the form won't get emptied if the valid-
ation fails.