Fix: confirm password field not listening for input changes in certain situations
If the user entered a password, then a confirm password,
then went back and edited the password, it would correctly
display an error for the confirm password field. However,
it would not listen for input events like in other situations.
This was because the listener would only be attached on blur
or on submit events.

So now an input event listener is attached to the confirm
password field if the password field is changed, and the
confirm password field is not empty, and fails validation,
i.e. does not match the value of the password field.