Fix: make validationFn callbacks explicitly return bool result
Previously the value was returned automatically when
invoked by an arrow function, but if the arrow function
uses curly brackets then you have to return manually.

Returned manually in all cases to make it apparent
what is going on.

Also made confirmPassword field validate when password
input is edited, or if confirmPassword is edited, after
confirmPassword blur has occurred or there is a non-blank
value in the confirmPassword input.