Encode passport messages as JSON so view knows which error is for which field
Passport only allows messages to be strings. Had a
hacky workaround so that the email form field could
have an error if the email/user does not exist, and
the password field an error if the user/password
does not match - but this is a better solution.

I think I just totally forgot JSON existed...