Improve accessibility by making form errors notify screen reader users
However this does not work with any server side errors
that are sent over as part of a page refresh/redirect.
So if you try to log in with a user that doesn't exist,
the error will be sent over as part of the HTML. As
the DOM doesn't see an update, the aria-live polite
error element does not announce anything.

Not sure how to fix this other than keeping the errors
in a js object and adding them to the html after it has
finished loading.

Surely there is a better way of doing this with some
built in functionality?